+++ src/engines/html5/media-source/adapters/native-adapter.js
@@ -10,6 +10,7 @@ import {getSuitableSourceForResolution} from '../../../../utils/resolution'
 import * as Utils from '../../../../utils/util'
 import FairPlay from '../../../../drm/fairplay'
 import Env from '../../../../utils/env'
+import FakeEvent from '../../../../event/fake-event'
 import PlayerError from "../../../../utils/player-error";
 
 /**
@@ -74,6 +75,12 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @private
    */
   _progressiveSources: Array<Source>;
+  /**
+   * The player tracks.
+   * @member {Array<Track>} - _playerTracks
+   * @private
+   */
+  _playerTracks: Array<Track>;
 
   /**
    * Checks if NativeAdapter can play a given mime type.
@@ -187,16 +194,8 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
 
     if (!this._loadPromise) {
       this._loadPromise = new Promise((resolve, reject) => {
-        // We're using 'loadeddata' event for native hls (on 'loadedmetadata' native hls doesn't have tracks yet).
-        this._eventManager.listenOnce(this._videoElement, Html5Events.LOADED_DATA, () => {
-          let data = {tracks: this._getParsedTracks()};
-          NativeAdapter._logger.debug('The source has been loaded successfully');
-          resolve(data);
-        });
-        this._eventManager.listenOnce(this._videoElement, Html5Events.ERROR, (error) => {
-          NativeAdapter._logger.error(error);
-          reject(error);
-        });
+        this._eventManager.listenOnce(this._videoElement, Html5Events.LOADED_DATA, this._onLoadedData.bind(this, resolve));
+        this._eventManager.listenOnce(this._videoElement, Html5Events.ERROR, this._onError.bind(this, reject));
         if (this._isProgressivePlayback()) {
           this._setProgressiveSource();
         }
@@ -211,6 +210,39 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
       });
     }
     return this._loadPromise;
+  }
+
+  /**
+   * Loaded data event handler.
+   * @param {Function} resolve - The resolve promise function.
+   * @private
+   * @returns {void}
+   */
+  _onLoadedData(resolve: Function): void {
+    const parseTracksAndResolve = () => {
+      this._playerTracks = this._getParsedTracks();
+      this._addNativeAudioTrackChangeListener();
+      this._addNativeTextTrackChangeListener();
+      NativeAdapter._logger.debug('The source has been loaded successfully');
+      resolve({tracks: this._playerTracks});
+    };
+    if (this._videoElement.textTracks.length > 0) {
+      parseTracksAndResolve();
+    } else {
+      this._eventManager.listenOnce(this._videoElement, Html5Events.CAN_PLAY, parseTracksAndResolve.bind(this));
+    }
+  }
+
+  /**
+   * Error event handler.
+   * @param {Function} reject - The reject promise function.
+   * @param {FakeEvent} error - The error fake event.
+   * @private
+   * @returns {void}
+   */
+  _onError(reject: Function, error: FakeEvent): void {
+    NativeAdapter._logger.error(error);
+    reject(error);
   }
 
   /**
@@ -238,9 +270,9 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @private
    */
   _getParsedTracks(): Array<Track> {
-    let videoTracks = this._getParsedVideoTracks();
-    let audioTracks = this._getParsedAudioTracks();
-    let textTracks = this._getParsedTextTracks();
+    const videoTracks = this._getParsedVideoTracks();
+    const audioTracks = this._getParsedAudioTracks();
+    const textTracks = this._getParsedTextTracks();
     return videoTracks.concat(audioTracks).concat(textTracks);
   }
 
@@ -265,11 +297,11 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @private
    */
   _getParsedProgressiveVideoTracks(): Array<Track> {
-    let videoTracks = this._progressiveSources;
-    let parsedTracks = [];
+    const videoTracks = this._progressiveSources;
+    const parsedTracks = [];
     if (videoTracks) {
       for (let i = 0; i < videoTracks.length; i++) {
-        let settings = {
+        const settings = {
           id: videoTracks[i].id,
           bandwidth: videoTracks[i].bandwidth,
           width: videoTracks[i].width,
@@ -291,11 +323,11 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    */
   _getParsedAdaptiveVideoTracks(): Array<Track> {
     //TODO check adaptation in safari hls
-    let videoTracks = this._videoElement.videoTracks;
-    let parsedTracks = [];
+    const videoTracks = this._videoElement.videoTracks;
+    const parsedTracks = [];
     if (videoTracks) {
       for (let i = 0; i < videoTracks.length; i++) {
-        let settings = {
+        const settings = {
           //TODO calculate width/height/bandwidth
           id: videoTracks[i].id,
           active: videoTracks[i].selected,
@@ -312,15 +344,15 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
   /**
    * Get the parsed audio tracks
    * @function _getParsedAudioTracks
-   * @returns {Array<Track>} - The parsed audio tracks
+   * @returns {Array<AudioTrack>} - The parsed audio tracks
    * @private
    */
-  _getParsedAudioTracks(): Array<Track> {
-    let audioTracks = this._videoElement.audioTracks;
-    let parsedTracks = [];
+  _getParsedAudioTracks(): Array<AudioTrack> {
+    const audioTracks = this._videoElement.audioTracks;
+    const parsedTracks = [];
     if (audioTracks) {
       for (let i = 0; i < audioTracks.length; i++) {
-        let settings = {
+        const settings = {
           id: audioTracks[i].id,
           active: audioTracks[i].enabled,
           label: audioTracks[i].label,
@@ -336,22 +368,24 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
   /**
    * Get the parsed text tracks
    * @function _getParsedTextTracks
-   * @returns {Array<Track>} - The parsed text tracks
+   * @returns {Array<PKTextTrack>} - The parsed text tracks
    * @private
    */
-  _getParsedTextTracks(): Array<Track> {
-    let textTracks = this._videoElement.textTracks;
-    let parsedTracks = [];
+  _getParsedTextTracks(): Array<PKTextTrack> {
+    const textTracks = this._videoElement.textTracks;
+    const parsedTracks = [];
     if (textTracks) {
       for (let i = 0; i < textTracks.length; i++) {
-        let settings = {
+        const settings = {
           kind: textTracks[i].kind,
           active: textTracks[i].mode === 'showing',
           label: textTracks[i].label,
           language: textTracks[i].language,
           index: i
         };
-        parsedTracks.push(new PKTextTrack(settings));
+        if (settings.language || settings.label) {
+          parsedTracks.push(new PKTextTrack(settings));
+        }
       }
     }
     return parsedTracks;
@@ -420,8 +454,10 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   selectAdaptiveVideoTrack(videoTrack: VideoTrack): void {
-    let videoTracks = this._videoElement.videoTracks;
-    if ((videoTrack instanceof VideoTrack) && videoTracks && videoTracks[videoTrack.index]) {
+    const videoTracks = this._videoElement.videoTracks;
+    if ((videoTrack instanceof VideoTrack)
+      && videoTracks
+      && videoTracks[videoTrack.index]) {
       this._disableVideoTracks();
       videoTracks[videoTrack.index].selected = true;
       this._onTrackChanged(videoTrack);
@@ -436,11 +472,69 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   selectAudioTrack(audioTrack: AudioTrack): void {
-    let audioTracks = this._videoElement.audioTracks;
-    if ((audioTrack instanceof AudioTrack) && audioTracks && audioTracks[audioTrack.index]) {
+    const audioTracks = this._videoElement.audioTracks;
+    if ((audioTrack instanceof AudioTrack)
+      && audioTracks
+      && audioTracks[audioTrack.index]) {
+      this._removeNativeAudioTrackChangeListener();
       this._disableAudioTracks();
       audioTracks[audioTrack.index].enabled = true;
       this._onTrackChanged(audioTrack);
+      this._addNativeAudioTrackChangeListener();
+    }
+  }
+
+  /**
+   * Remove the onchange listenr of the video element AudioTrackList.
+   * @private
+   * @returns {void}
+   */
+  _removeNativeAudioTrackChangeListener(): void {
+    if (this._videoElement.audioTracks) {
+      this._eventManager.unlisten(this._videoElement.audioTracks, 'change');
+    }
+  }
+
+  /**
+   * Add the onchange listenr of the video element AudioTrackList.
+   * @private
+   * @returns {void}
+   */
+  _addNativeAudioTrackChangeListener(): void {
+    if (this._videoElement.audioTracks) {
+      this._eventManager.listen(this._videoElement.audioTracks, 'change', () => this._onNativeAudioTrackChange());
+    }
+  }
+
+  /**
+   * Handler of the video element AudioTrackList onchange event.
+   * @private
+   * @returns {void}
+   */
+  _onNativeAudioTrackChange(): void {
+    const pkAudioTracks = this._playerTracks.filter(track => track instanceof AudioTrack);
+    const getActivePKAudioTrackIndex = () => {
+      const activeAudioTrack = pkAudioTracks.find(track => track.active === true);
+      return (activeAudioTrack ? activeAudioTrack.index : -1);
+    };
+    const getActiveVidAudioTrackIndex = () => {
+      for (let i = 0; i < this._videoElement.audioTracks.length; i++) {
+        const audioTrack = this._videoElement.audioTracks[i];
+        if (audioTrack.enabled) {
+          return i;
+        }
+      }
+      return -1;
+    };
+    NativeAdapter._logger.debug('Video element audio track change');
+    const vidIndex = getActiveVidAudioTrackIndex();
+    const pkIndex = getActivePKAudioTrackIndex();
+    if (vidIndex !== pkIndex) {
+      const pkAudioTrack = pkAudioTracks.find(track => track.index === vidIndex);
+      if (pkAudioTrack) {
+        NativeAdapter._logger.debug('Native selection of track, update the player audio track (' + pkIndex + ' -> ' + vidIndex + ')');
+        this._onTrackChanged(pkAudioTrack);
+      }
     }
   }
 
@@ -452,12 +546,83 @@ export default class NativeAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   selectTextTrack(textTrack: PKTextTrack): void {
-    let textTracks = this._videoElement.textTracks;
-    if ((textTrack instanceof PKTextTrack) && (textTrack.kind === 'subtitles' || textTrack.kind === 'captions') && textTracks && textTracks[textTrack.index]) {
+    const textTracks = this._videoElement.textTracks;
+    if ((textTrack instanceof PKTextTrack)
+      && (textTrack.kind === 'subtitles' || textTrack.kind === 'captions')
+      && textTracks && textTracks[textTrack.index]) {
+      this._removeNativeTextTrackChangeListener();
       this._disableTextTracks();
       textTracks[textTrack.index].mode = 'hidden';
       NativeAdapter._logger.debug('Text track changed', textTrack);
       this._onTrackChanged(textTrack);
+      this._addNativeTextTrackChangeListener();
+    }
+  }
+
+  /**
+   * Remove the onchange listenr of the video element TextTrackList.
+   * @private
+   * @returns {void}
+   */
+  _removeNativeTextTrackChangeListener(): void {
+    if (this._videoElement.textTracks) {
+      this._eventManager.unlisten(this._videoElement.textTracks, 'change');
+    }
+  }
+
+  /**
+   * Add the onchange listenr of the video element TextTrackList.
+   * @private
+   * @returns {void}
+   */
+  _addNativeTextTrackChangeListener(): void {
+    if (this._videoElement.textTracks) {
+      this._eventManager.listen(this._videoElement.textTracks, 'change', () => this._onNativeTextTrackChange());
+    }
+  }
+
+  /**
+   * Handler of the video element TextTrackList onchange event.
+   * @private
+   * @returns {void}
+   */
+  _onNativeTextTrackChange(): void {
+    const pkTextTracks = this._playerTracks.filter(track => track instanceof PKTextTrack);
+    const pkOffTrack = pkTextTracks.find(track => track.language === 'off');
+    const getActivePKTextTrackIndex = () => {
+      const activeTextTrack = pkTextTracks.find(track => track.active === true);
+      return (activeTextTrack ? activeTextTrack.index : -1);
+    };
+    const getActiveVidTextTrackIndex = () => {
+      for (let i = 0; i < this._videoElement.textTracks.length; i++) {
+        const textTrack = this._videoElement.textTracks[i];
+        if (textTrack.mode === 'showing') {
+          return i;
+        }
+      }
+      return -1;
+    };
+    NativeAdapter._logger.debug('Video element text track change');
+    const vidIndex = getActiveVidTextTrackIndex();
+    const pkIndex = getActivePKTextTrackIndex();
+    if (vidIndex !== pkIndex) {
+      // In case no text track with 'showing' mode
+      // we need to set the off track
+      if (vidIndex == -1) {
+        if (pkOffTrack) {
+          NativeAdapter._logger.debug('Native selection of track, update the player text track (' + pkIndex + ' -> off)');
+          this._onTrackChanged(pkOffTrack);
+        }
+      } else {
+        // In case the text track on the video element is
+        // different then the text track of the player
+        // we need to set the correct one
+        const pkTextTrack = pkTextTracks.find(track => track.index === vidIndex);
+        if (pkTextTrack) {
+          NativeAdapter._logger.debug('Native selection of track, update the player text track (' + pkIndex + ' -> ' + vidIndex + ')');
+          this._onTrackChanged(pkTextTrack);
+        }
+      }
     }
   }