+++ test/src/event/events.spec.js
@@ -28,6 +28,12 @@ describe('events', () => {
       WAITING: 'waiting',
     });
     events.CUSTOM_EVENTS.should.deep.equals({
+      ENTER_FULLSCREEN: 'enterfullscreen',
+      EXIT_FULLSCREEN: 'exitfullscreen',
+      REQUESTED_ENTER_FULLSCREEN: 'requestedenterfullscreen',
+      REQUESTED_EXIT_FULLSCREEN: 'requestedexitfullscreen',
+      AUTOPLAY_FAILED: 'autoplayfailed',
+      FALLBACK_TO_MUTED_AUTOPLAY: 'fallbacktomutedautoplay',
       CHANGE_SOURCE_STARTED: 'changesourcestarted',
       CHANGE_SOURCE_ENDED: 'changesourceended',
       MUTE_CHANGE: 'mutechange',